@@ -1,64 +1,48 @@
 {
-  "name": "now-ui-react-native",
-  "version": "1.3.0",
-  "description": "Essence global mart",
   "main": "node_modules/expo/AppEntry.js",
+  "description": "Essence global mart",
   "scripts": {
     "start": "expo start",
     "android": "expo start --android",
     "ios": "expo start --ios",
-    "eject": "expo eject",
-    "lint": "eslint *.js **/*.js"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/creativetimofficial/now-ui-react-native.git"
+    "web": "expo start --web",
+    "eject": "expo eject"
   },
   "dependencies": {
-    "@react-native-community/masked-view": "0.1.10",
-    "@react-navigation/bottom-tabs": "^5.5.1",
-    "@react-navigation/compat": "^5.1.25",
-    "@react-navigation/drawer": "5.8.1",
-    "@react-navigation/native": "5.5.0",
-    "@react-navigation/stack": "5.4.1",
-    "expo": "^39.0.0",
-    "expo-asset": "~8.2.0",
-    "expo-font": "~8.3.0",
+    "@react-native-async-storage/async-storage": "^1.14.1",
+    "@react-native-community/masked-view": "^0.1.10",
+    "@react-navigation/bottom-tabs": "^5.11.8",
+    "@react-navigation/compat": "^5.3.14",
+    "@react-navigation/drawer": "^5.12.4",
+    "@react-navigation/native": "^5.9.3",
+    "@react-navigation/stack": "^5.14.3",
+    "axios": "^0.21.1",
+    "constate": "^3.2.0",
+    "expo": "~40.0.0",
+    "expo-app-loading": "^1.0.1",
+    "expo-asset": "^8.2.2",
+    "expo-font": "^8.4.0",
+    "expo-status-bar": "~1.0.3",
+    "firebase": "^7.9.0",
     "galio-framework": "^0.7.1",
+    "lottie-react-native": "^3.5.0",
     "prop-types": "^15.7.2",
-    "react": "16.13.1",
-    "react-native": "https://github.com/expo/react-native/archive/sdk-39.0.3.tar.gz",
-    "react-native-gesture-handler": "~1.7.0",
-    "react-native-modal-dropdown": "git+https://github.com/siemiatj/react-native-modal-dropdown.git",
-    "react-native-reanimated": "~1.13.0",
-    "react-native-svg": "12.1.0",
-    "react-native-safe-area-context": "3.1.4",
-    "react-native-screens": "~2.10.1"
+    "react": "^17.0.1",
+    "react-dom": "16.13.1",
+    "react-native": "^0.64.0",
+    "react-native-animated-loader": "^0.0.9",
+    "react-native-gesture-handler": "^1.8.0",
+    "react-native-modal-dropdown": "^1.0.0",
+    "react-native-reanimated": "1.13.2",
+    "react-native-safe-area-context": "3.1.9",
+    "react-native-safe-area-view": "^1.1.1",
+    "react-native-screens": "^2.15.2",
+    "react-native-svg": "^12.1.0",
+    "react-native-web": "~0.13.12",
+    "react-native-webview": "11.0.0"
   },
   "devDependencies": {
-    "babel-eslint": "^10.0.3",
-    "babel-preset-expo": "^8.3.0"
+    "@babel/core": "~7.9.0"
   },
-  "keywords": [
-    "Essence-Mart",
-    "Essence global Mart system",
-    "Essence global Mart",
-    "now iOS",
-    "react native iOS",
-    "creative tim",
-    "now Android",
-    "react native ui kit",
-    "react native expo",
-    "freebie",
-    "react native now design",
-    "react native galio",
-    "galio now pro",
-    "galio react native free app",
-    "now expo react native",
-    "react native ui template"
-  ],
-  "author": "Akolade, Efosa",
-  "bugs": {
-    "url": "https://github.com/Efosaok/Mobile-App-Essence-Mart/issues"
-  }
+  "private": true
 }
