@@ -1,12 +1,12 @@
 const COLORS = {
   DEFAULT: '#888888',
-  PRIMARY: '#f96332',
+  PRIMARY: '#AD1457',
   LABEL: '#FE2472',
   INFO: '#2CA8FF',
   ERROR: '#FF3636',
   SUCCESS: '#18ce0f',
   WARNING: '#FFB236',
-  SECONDARY: '#444444',
+  SECONDARY: '#F06292',
   NEUTRAL: 'rgba(255, 255, 255, 0.2)',
   TABS: 'rgba(222, 222, 222, 0.3)',
   TEXT: '#32325D',
